movie_back/requirements.txt
Django==4.2.19
djangorestframework==3.15.2
django-cors-headers==4.7.0
Pillow==11.1.0
